src/main/java/com/kakaopay/membership/domain/member/MemberWriter.java
package com.kakaopay.membership.domain.member;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Component;
import org.springframework.transaction.annotation.Propagation;
import org.springframework.transaction.annotation.Transactional;

@RequiredArgsConstructor
@Component
public class MemberWriter {

    private final MemberRepository memberRepository;

    @Transactional(propagation = Propagation.REQUIRES_NEW)
    public Member write(Long memberId) {
        Member member = Member.create(memberId);
        return memberRepository.save(member);
    }

}
